rename: 주소 정보 파일 구조 수정 (#5)
