+++ src/main/java/com/newfit/reservation/repository/EmdAreaRepository.java
@@ -1,6 +1,6 @@
 package com.newfit.reservation.repository;
 
-import com.newfit.reservation.domain.EmdArea;
+import com.newfit.reservation.domain.location.EmdArea;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
