+++ src/main/java/com/newfit/reservation/repository/SiggAreaRepository.java
@@ -1,6 +1,6 @@
 package com.newfit.reservation.repository;
 
-import com.newfit.reservation.domain.SiggArea;
+import com.newfit.reservation.domain.location.SiggArea;
 import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
